Separate into multiple files
Still need to fix some destructors.
